fix: don't recommend an implementation for Linux yet
We specified unifiedpush_linux as the default platform implementation
for Linux, but right now we shouldn't recommend anything yet. If people
want to support Linux they can pull it in themselves for now.

Resolves the following warning:
Package unifiedpush:linux references unifiedpush_linux:linux as the default plugin, but the package does not exist, or is not a plugin package.
Ask the maintainers of unifiedpush to either avoid referencing a default implementation via `platforms: linux: default_package: unifiedpush_linux` or create a plugin named
unifiedpush_linux.